Added resuable code
